Add notes for local docker image
This is how I am running on my local machine, should be usable in the field. 

I am on Ubuntu through "WSL", windows subsystem. Note that the docker install for WSL is a bit weird, it works best to do the Windows install and that will naturally recognize that you are using WSL.